+++ example/enumMap/src/main/java/io/github/hanbernate/jsonbom/example/OrderLogRepository.java
@@ -6,7 +6,6 @@ import io.github.hanbernate.jsonbom.api.Bom;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.concurrent.TimeUnit;
 import lombok.Data;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;