+++ example/queryDbOnDemand/src/main/java/io/github/hanbernate/jsonbom/spring/AbstractBomDao.java
@@ -11,7 +11,6 @@ import java.util.Optional;
 import java.util.Map.Entry;
 import java.util.stream.Collectors;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;