+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ## [Unreleased]
 
 ### 0.2.0.rc2
-- 修复BomPropertyDefinitionProvider在处理泛型时异常
+- 修复BomPropertyDefinitionProvider在处理泛型时异常，并支持@BomMapping注解；已知问题：全局注册的valueHandler无法识别，会影响mcp的toollist输出；mcp场景下请使用注解方式设置valueNode=true
 ### 0.2.0.rc1
 - 新增MCP支持。使用 BomPropertyDefinitionProvider，配合 victools jsonschema-generator 自动生成带 @BomType 注解字段的 JSON Schema。如何与spring-ai集成可以参考example/mcpServer模块
 ### 0.1.0